Feature HTTPS Client PUT and POST support (#1170)
* Add support for PUT/POST HTTP methods.

* Add S3 Synchronous and Asynchronous Upload Demo that utilizes PUT.

* Add system integration tests for PUT and POST methods.

* Fix signalling that a response is complete for both sync and async.
  The semaphore saying the synchronous response is complete and the asynchronous callback 
  signalling the response is complete must be called mutually exclusively otherwise a modified
  buffer during a context switch may  cause them both to be called.

* Update IDE projects for HTTPS Client upload demos.

* Force the use of SigV4 in the presigned url generation script.
  SigV2 is used by default in the version 1.9.182 of boto3. SigV2 is being deprecated.
  No matter what version the user's Python development has, this commit forces the use of
  SigV4.

* Make the HTTPS Client send sync timeout configurable.
  The send sync timeout in the demos was 0 which was a max blocking.
  S3 allows only 100 requests max per connection. The docs have been updated to note this.

+++ libraries/c_sdk/standard/https/test/unit/iot_tests_https_client.c
@@ -63,7 +63,7 @@
     "Vary: *\r\n"                              \
     "P3P: CP=\"This is not a P3P policy\"\r\n" \
     "xserver: www1021\r\n"
-#define HTTPS_TEST_RESPONSE_HEADER_LINES_LENGTH    sizeof( HTTPS_TEST_RESPONSE_HEADER_LINES ) - 1
+#define HTTPS_TEST_RESPONSE_HEADER_LINES_LENGTH    sizeof( HTTPS_TEST_RESPONSE_HEADER_LINES ) - 1 /**< @brief The length of the HTTP response test header lines. */
 
 /**
  * @brief Header lines with no content-length for testing.
@@ -75,23 +75,23 @@
     "Vary: *\r\n"                                          \
     "P3P: CP=\"This is not a P3P policy\"\r\n"             \
     "xserver: www1021\r\n\r\n"
-#define HTTPS_TEST_RESPONSE_HEADER_LINES_NO_CONTENT_LENGTH_LENGTH    sizeof( HTTPS_TEST_RESPONSE_HEADER_LINES_NO_CONTENT_LENGTH ) - 1
+#define HTTPS_TEST_RESPONSE_HEADER_LINES_NO_CONTENT_LENGTH_LENGTH    sizeof( HTTPS_TEST_RESPONSE_HEADER_LINES_NO_CONTENT_LENGTH ) - 1 /**< @brief Length of the HTTP test response headers where there is no Content-Length. */
 
 /**
- * @brief Header name and values to verify reading the header.
+ * Header name and values to verify reading the header.
  */
-#define HTTPS_DATE_HEADER                                            "Date"
-#define HTTPS_ETAG_HEADER                                            "ETag"
-#define HTTPS_NONEXISTENT_HEADER                                     "Non-Existent-Header"
-#define HTTPS_DATE_HEADER_VALUE                                      "Sun, 14 Jul 2019 06:07:52 GMT"
-#define HTTPS_ETAG_HEADER_VALUE                                      "\"3356-5233\""
-#define HTTPS_CONTENT_LENGTH_VALUE                                   ( 43 )
+#define HTTPS_DATE_HEADER                                            "Date"                          /**< @brief "Date" HTTP header field name. */
+#define HTTPS_ETAG_HEADER                                            "ETag"                          /**< @brief "ETag" HTTP header field name. */
+#define HTTPS_NONEXISTENT_HEADER                                     "Non-Existent-Header"           /**< @brief HTTP header field name of a non-existing header for testing. */
+#define HTTPS_DATE_HEADER_VALUE                                      "Sun, 14 Jul 2019 06:07:52 GMT" /**< @brief Test header value for the "Date" header field. */
+#define HTTPS_ETAG_HEADER_VALUE                                      "\"3356-5233\""                 /**< @brief Test header value for the "ETag" header field. */
+#define HTTPS_CONTENT_LENGTH_VALUE                                   ( ( uint32_t ) 43 )             /**< @brief Test header value as a unsigned integer for the "Content-Length" header field. */
 
 /**
- * @brief The array lengths of to store the header values.
+ * The array lengths of local value buffers in the test to store the header values.
  */
-#define HTTPS_TEST_VALUE_BUFFER_LENGTH_LARGE_ENOUGH                  ( 64 )
-#define HTTPS_TEST_VALUE_BUFFER_LENGTH_TOO_SMALL                     ( 8 )
+#define HTTPS_TEST_VALUE_BUFFER_LENGTH_LARGE_ENOUGH                  ( 64 ) /**< @brief A large enough test length of a local value buffer to store the returned header value. */
+#define HTTPS_TEST_VALUE_BUFFER_LENGTH_TOO_SMALL                     ( 8 )  /**< @brief A too small test length of a local value buffer to store the returned header value. */
 
 /*-----------------------------------------------------------*/
